报价成交 订单提交代码完善

+++ sj_api/src/main/java/cc/mrbird/febs/api/entity/SOfferPrice.java
@@ -39,7 +39,7 @@ public class SOfferPrice implements Serializable {
     /**
      * 任务转让主键
      */
-    private Integer taskOrderId;
+    private Long taskOrderId;
 
     /**
      * 竞标人
